Moved components into header/
